+++ src/ua/data_types/variant.rs
@@ -103,7 +103,21 @@ impl Variant {
     #[must_use]
     pub fn to_array<T: DataType>(&self) -> Option<ua::Array<T>> {
         if !unsafe { UA_Variant_hasArrayType(self.as_ptr(), T::data_type()) } {
-            return None;
+            // Special case: open62541 automatically converts arrays of extension objects into the
+            // contained data type (as of version 1.4). This only works for non-empty arrays since
+            // the element type needs to be known.
+            //
+            // To make handling such arrays easier in user code, we allow _coercion_ of such empty
+            // arrays into any data type.
+            let is_empty_structured_array = self.0.arrayLength == 0
+                && unsafe {
+                    UA_Variant_hasArrayType(self.as_ptr(), ua::ExtensionObject::data_type())
+                };
+            if !is_empty_structured_array {
+                return None;
+            }
+            // Fall through to let `ua::Array::from_raw_parts()` handle the distinction between an
+            // empty and an invalid array (where `self.0.data` is the sentinel value or null).
         }
         ua::Array::from_raw_parts(self.0.data.cast::<T::Inner>(), self.0.arrayLength)
     }