+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Coerce _empty_ arrays of `ua::ExtensionObject` into the requested data type. This mirrors the
+  auto-unwrapping behavior of open62541 version 1.4.
+
 ## [0.6.0-pre.3] - 2024-05-18
 
 [0.6.0-pre.3]: https://github.com/HMIProject/open62541/compare/v0.6.0-pre.2...v0.6.0-pre.3